pyproject.toml
[project]
name = "rozetka_keepa"
version = "0"
description = "Rozetka.ua Keepa"
authors = [
    { name = "Alexey ALERT Rubasheff", email = "alexey.rubasheff@gmail.com" }
]
readme = "README.md"
requires-python = "==3.12.*"
keywords = ["rozetka"]
classifiers = [
    "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = [
    "aiogram>=3.18.0",
    "apscheduler>=3.11.0",
    "global-logger>=0.4.3",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "rozetka-api>=1.6.9",
    "sqlalchemy>=2.0.38",
    "sqlalchemy-utils>=0.41.2",
    "tzdata>=2025.1",
]
[tool.uv]
prerelease = "allow"

[project.urls]
homepage = "https://github.com/ALERTua/rozetka_keepa"
repository = "https://github.com/ALERTua/rozetka_keepa"
documentation = "https://github.com/ALERTua/rozetka_keepa"
"Bug Tracker" = "https://github.com/ALERTua/rozetka_keepa/issues"
"Rozetka API" = "https://github.com/ALERTua/rozetka_api"
"Rozetka Keepa" = "https://github.com/ALERTua/rozetka_keepa"
"Rozetka FastAPI" = "https://github.com/ALERTua/rozetka_fastapi"
"Rozetka UserScript" = "https://github.com/ALERTua/rozetka_userscript"

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
    "pytest>=8.3.5",
    "ruff>=0.9.9",
]
